scoring/requirements.txt
pyyaml>=6.0
jiwer>=3.0
requests>=2.28
python-dotenv>=1.0
